split out encyrptied data hashing
