+++ test/test.c
@@ -135,7 +135,9 @@ int main(void) {
   test_rc_libretro(); /* libretro extensions require hash support */
   test_cdreader();
   test_hash();
+ #ifndef RC_HASH_NO_ZIP
   test_hash_zip();
+ #endif
 #endif
 #endif
